Makefile and friends
All necessary info to have a proper github site
and build system.

+++ Makefile
@@ -1,20 +1,50 @@
 PYTHON_3?=python3
 
-BUILDDIR = _site
+BUILDDIR = build
 POSTSDIR = posts
-CSSDIR   = css
+CACHEDIR = cache
 POSTS= $(addprefix $(BUILDDIR)/, $(patsubst %.md,%.html, $(wildcard $(POSTSDIR)/*.md)))
+CSS = $(addprefix $(BUILDDIR)/, $(wildcard css/*.css))
 FILTERS = $(wildcard filters/*.lua)
-DIRS = $(addprefix $(BUILDDIR)/, $(POSTSDIR))
+DIRS = $(addprefix $(BUILDDIR)/, $(POSTSDIR) css data img) $(CACHEDIR)
+TEMPLATES = $(wildcard templates/*)
+STATIC_FILES = $(addprefix $(BUILDDIR)/, $(wildcard data/*) $(wildcard img/*))
+
 
 .DEFAULT: all
-.PHONY: all clean server filters
+.PHONY: all clean server pages
 
-$(POSTS): $(BUILDDIR)/%.html : %.md $(FILTERS) | $(DIRS)
-	pandoc --lua-filter filters/tikz.lua --self-contained -s -f markdown -t html5 -o "$@" "$<"
+all: pages $(DIRS)
+
+pages: $(BUILDDIR)/posts.html $(BUILDDIR)/index.html $(BUILDDIR)/projects.html $(DIRS) $(POSTS) $(STATIC_FILES)
+
+$(POSTS): $(BUILDDIR)/%.html : %.md $(FILTERS) $(CSS) $(DIRS) $(TEMPLATES)
+	pandoc --metadata-file=config.yaml \
+	       --lua-filter filters/tikz.lua \
+	       --template=templates/default.html \
+               --katex -s -f markdown -t html5 -o "$@" "$<"
+
+$(BUILDDIR)/index.html: index.html $(CSS) $(DIRS) $(TEMPLATES)
+	pandoc --metadata-file=config.yaml \
+	       --template=templates/default.html \
+               -s -f html -t html5 -o "$@" "$<"
+
+$(BUILDDIR)/projects.html: projects.md $(CSS) $(DIRS) $(TEMPLATES)
+	pandoc --metadata-file=config.yaml \
+	       --template=templates/default.html \
+               -s -f markdown -t html5 -o "$@" "$<"
+
+$(BUILDDIR)/posts.html: posts.md $(CSS) $(DIRS) $(TEMPLATES)
+	pandoc --metadata-file=config.yaml \
+	       --template=templates/default.html \
+               -s -f markdown -t html5 -o "$@" "$<"
 
 
-all: $(POSTS) $(DIRS)
+$(CSS): $(BUILDDIR)/css/%.css : css/%.css $(DIRS)
+	cp "$<" "$@"
+
+$(STATIC_FILES): $(BUILDDIR)/% : % $(DIRS)
+	cp "$<" "$@"
 
 $(DIRS):
 	mkdir -p "$@"
@@ -24,11 +54,5 @@ server: ## Preview site
 	cd $(BUILDDIR) && $(PYTHON_3) -m http.server
 
 clean: ## Delete cache and build files
-	rm -r $(BUILDDIR)
-	rm -r _cache
-
-
-help:
-	@grep -E '^[a-zA-Z0-9_-]+:.*?## .*$$' $(MAKEFILE_LIST) \
-	| sed -n 's/^\(.*\): \(.*\)##\(.*\)/\1\3/p' \
-	| column -t  -s ' '
+	if [ -d $(BUILDDIR) ]; then rm -r $(BUILDDIR); fi
+	if [ -d $(CACHEDIR) ]; then rm -r $(CACHEDIR); fi
